// github/types/paths/repos/{owner}/{repo}/releases/{release_id}.types.ts
export type reposGetRelease = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: reposGetRelease_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: release;
          };
        };
        examples: {
          default: unknown;
        };
      };
      401: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Users with push access to the repository can edit a release.
 */
export type reposUpdateRelease = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: reposUpdateRelease_Path;
    headers: never;
    cookie: never;
    body: {
      /**
       * The name of the tag.
       */
      tag_name?: string;
      /**
       * Specifies the commitish value that determines where the Git tag is created from. Can be any branch or commit SHA. Unused if the Git tag already exists. Default: the repository's default branch.
       */
      target_commitish?: string;
      /**
       * The name of the release.
       */
      name?: string;
      /**
       * Text describing the contents of the tag.
       */
      body?: string;
      /**
       * `true` makes the release a draft, and `false` publishes the release.
       */
      draft?: boolean;
      /**
       * `true` to identify the release as a prerelease, `false` to identify the release as a full release.
       */
      prerelease?: boolean;
      /**
       * Specifies whether this release should be set as the latest release for the repository. Drafts and prereleases cannot be set as latest. Defaults to `true` for newly published releases. `legacy` specifies that the latest release should be determined based on the release creation date and higher semantic version.
       * @default "true"
       */
      make_latest?: "true" | "false" | "legacy";
      /**
       * If specified, a discussion of the specified category is created and linked to the release. The value must be a category that already exists in the repository. If there is already a discussion linked to the release, this parameter is ignored. For more information, see "[Managing categories for discussions in your repository](https://docs.github.com/discussions/managing-discussions-for-your-community/managing-categories-for-discussions-in-your-repository)."
       */
      discussion_category_name?: string;
    };
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: release;
          };
        };
        examples: {
          default: unknown;
        };
      };
      404: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: basic_error;
          };
        };
        examples: {};
      };
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

/**
 * Users with push access to the repository can delete a release.
 */
export type reposDeleteRelease = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: reposDeleteRelease_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type reposGetRelease_Path = {
  owner: string;
  repo: string;
  release_id: number;
};

export type reposUpdateRelease_Path = {
  owner: string;
  repo: string;
  release_id: number;
};

export type reposDeleteRelease_Path = {
  owner: string;
  repo: string;
  release_id: number;
};